Improve the setup instructions [ci skip]
Before being able to run the migrations we need to make sure to have
postgres running and that we have created a database. Running "rake
db:setup" is a shortcut for "rake db:create; rake db:schema:load, and
rake db:seed".

We also need to make sure that we set up the test environment correctly
otherwise we'll run into errors when running the specs.